+++ Sources/Swift/Integrations/SessionReplay/SentryBreadcrumbReplayConverter.swift
@@ -76,8 +76,7 @@ class SentryBreadcrumbReplayConverter: NSObject {
         return SentryRRWebSpanEvent(timestamp: timestamp, endTimestamp: timestamp, operation: "resource.http", description: description, data: data)
     }
     
-    private  func getLevel(breadcrumb: Breadcrumb) -> SentryLevel {
-        return SentryLevel(rawValue: SentryLevelHelper.breadcrumbLevel(breadcrumb)) ?? .none
-        
+    private func getLevel(breadcrumb: Breadcrumb) -> String {
+        return SentryLevelHelper.breadcrumbLevel(breadcrumb)
     }
 }