+++ Sources/Swift/Integrations/SessionReplay/SentrySRDefaultBreadcrumbConverter.swift
@@ -83,7 +83,7 @@ class SentrySRDefaultBreadcrumbConverter: NSObject, SentryReplayBreadcrumbConver
         return SentryRRWebSpanEvent(timestamp: startTimestamp, endTimestamp: timestamp, operation: "resource.http", description: description, data: data)
     }
     
-    private func getLevel(breadcrumb: Breadcrumb) -> SentryLevel {
-        return SentryLevel(rawValue: SentryLevelHelper.breadcrumbLevel(breadcrumb)) ?? .none
+    private func getLevel(breadcrumb: Breadcrumb) -> String {
+        return SentryLevelHelper.breadcrumbLevel(breadcrumb)
     }
 }